Refactor nav bar to make nav-bar a server component, and nav-links and nav-bar-mobile client components
