+++ frontend/src/components/layout/nav-bar-mobile.tsx
@@ -1,20 +1,21 @@
-
-
+"use client";
 import Link from "next/link";
 import { Button, Menu } from "@mantine/core";
 import { IconMenu2, IconSearch } from "@tabler/icons-react";
 import Logo from "@/components/layout/logo";
 import SearchBar from "@/components/search/search-bar";
+import { useState } from "react";
+import { usePathname } from "next/navigation";
 
-export const NavBarMobile = ({
-  showSearch,
-  setShowSearch,
-  pathname,
-}: {
-  showSearch: boolean;
-  setShowSearch: (show: boolean) => void;
-  pathname: string;
-}) => {
+export const NavBarMobile = () => {
+  const [showSearch, setShowSearch] = useState(false);
+  const pathname = usePathname();
+
+  const menuItems = [
+    { href: "/", label: "Home" },
+    { href: "/jobs", label: "Jobs" },
+  ];
+
   return (
     <>
       {showSearch ? (
@@ -48,20 +49,16 @@ export const NavBarMobile = ({
                 </Button>
               </Menu.Target>
               <Menu.Dropdown>
-                <Menu.Item
-                  component={Link}
-                  href="/"
-                  className={pathname === "/" ? "font-bold" : ""}
-                >
-                  Home
-                </Menu.Item>
-                <Menu.Item
-                  component={Link}
-                  href="/jobs"
-                  className={pathname === "/jobs" ? "font-bold" : ""}
-                >
-                  Jobs
-                </Menu.Item>
+                {menuItems.map((item) => (
+                  <Menu.Item
+                    key={item.href}
+                    component={Link}
+                    href={item.href}
+                    className={pathname === item.href ? "font-bold" : ""}
+                  >
+                    {item.label}
+                  </Menu.Item>
+                ))}
               </Menu.Dropdown>
             </Menu>
           </div>
